Added data set, tester object, and data manipulation object.
Added parameters and multiple functions for the neural network class. Created tester class
to handle testing a neural network and remove some of the hastle of managing the data sets.
You can also just use the data object to handle training the model without using the tester.
I also added the data set in a zip file to the project. This will allow others to download and
experiment with it themselves.